Fix race in the command-response process
One of the tests randomly broke because of a race in the
command-response processing. Sometimes the handleCommandResponse helper didn't
drop a duplicated response (i.e. having the same sequence number than
the previous one) if Exec() wasn't "fast enough" to increase the
counter after getting the response. The duplicate then successfully made
it to the single response channel (cmds) and the next command got this
back.

The fix introduces dedicated response channels per sequence number
stored in a map. When we get a response, its sequence number is looked
up in the map and if there is a channel (opened previously by Exec())
the response is sent to that channel. Then the channel is closed and
removed from the map immediately. If there is no channel found for the
response then it will be dropped.